Move generate_unique_id() from Banner to GlobalTool
