+++ utility/global_tool.py
@@ -8,6 +8,8 @@ Author : DrLarck
 Last update : 07/04/20 by DrLarck
 """
 
+from string import ascii_letters
+
 
 class GlobalTool:
 
@@ -49,3 +51,42 @@ class GlobalTool:
             rarity = 5
 
         return rarity
+
+    @staticmethod
+    async def generate_unique_id(reference):
+        """
+        Generate a unique id from the reference
+
+        :param reference: (`int`)
+
+        --
+
+        :return: `str`
+        """
+
+        # Init
+        letters = ascii_letters
+
+        # Generation
+        # First of all, store the highest value in 'number'
+        number = int(reference / pow(52, 4))
+        reference -= number * pow(52, 4)
+
+        # Then deal the value with the letters
+        # Each letter can store (52^index - 1) values
+        # The first_letter (tier 1) can handle 52 values
+        first_letter = int(reference / pow(52, 3))
+        reference -= first_letter * pow(52, 3)
+
+        second_letter = int(reference / pow(52, 2))
+        reference -= second_letter * pow(52, 2)
+
+        third_letter = int(reference / 52)
+        reference -= third_letter * 52
+
+        fourth_letter = reference
+
+        # Get the unique id
+        id_ = f"{letters[fourth_letter]}{letters[third_letter]}{letters[second_letter]}{letters[first_letter]}{number}"
+
+        return id_
